tiny refactor, with the intention to add support for other issue providers (eg bitbucket, gitlab, phabricator)
